# backend/swagger.yml
openapi: 3.0.3
info:
  title: Kids Jobs Backend API
  description: |
    API standalone do kids-jobs para vagas, fontes, scrapings
    e operacao local.
  version: 2.0.0
servers:
  - url: http://localhost:8001
    description: Local development
...
paths:
...
  /api/v1/scrapers/config:
    get:
      tags: [scrapers]
      summary: Get full scraping config tree
      operationId: getFullConfig
      description: |
        Returns Sources -> Categories tree.
        Used by the scheduler to build scraping jobs.
      parameters:
        - name: enabled_only
          in: query
          schema: { type: boolean, default: true }
      responses:
        "200":
          description: Scraping config tree
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SourceConfig" }
...
  /api/v1/scrapers/run-all:
    post:
      tags: [scrapers]
      summary: Queue manual execution for all enabled scrapers
      operationId: runAllScrapers
      responses:
        "202":
          description: Batch execution queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScraperRunQueuedResponse" }
        "400":
          description: No enabled source available
  /api/v1/scrapers/{source_id}/run:
    post:
      tags: [scrapers]
      summary: Queue manual scraper execution for a source
      operationId: runScraper
      parameters:
        - name: source_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "202":
          description: Scraper execution queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScraperRunQueuedResponse" }
        "400":
          description: Source is disabled
        "404":
          description: Source not found
        "503":
          description: Manual executor unavailable

  # --- Sources --------------------------------------------------------------
...
  /api/v1/sources/{source_id}/categories:
    get:
      tags: [categories]
      summary: List categories by source
      operationId: listCategories
      parameters:
        - name: source_id
          in: path
          required: true
          schema: { type: integer }
        - name: enabled_only
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Category" }
        "404":
          description: Source not found
    post:
      tags: [categories]
      summary: Create category for source
      operationId: createCategory
      parameters:
        - name: source_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "404":
          description: Source not found
  /api/v1/categories/{category_id}:
    get:
      tags: [categories]
      summary: Get category by ID
      operationId: getCategory
      parameters:
        - name: category_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Category details
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "404":
          description: Not found
    put:
      tags: [categories]
      summary: Update category
      operationId: updateCategory
      parameters:
        - name: category_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "404":
          description: Not found
    delete:
      tags: [categories]
      summary: Delete category
      operationId: deleteCategory
      parameters:
        - name: category_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
components:
  schemas:
    ScraperRunQueuedResponse:
      type: object
      properties:
        job_id: { type: string }
        status: { type: string, example: queued }
        queued_count: { type: integer }
        submitted_at: { type: string, format: date-time }
        source_id: { type: integer, nullable: true }
        source_name: { type: string, nullable: true }
    SourceConfig:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        enabled: { type: boolean }
        scraper_base_url: { type: string }
        scraper_type: { type: string }
        scraper_schedule: { type: string }
        analysis: { type: string }
        description: { type: string }
        categories:
          type: array
          items: { $ref: "#/components/schemas/CategoryConfig" }
    CategoryConfig:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        scrape_path: { type: string }
        category_id: { type: integer, nullable: true }
        schedule: { type: string }
        enabled: { type: boolean }
...
    Category:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        source_id: { type: integer }
        category_id: { type: integer, nullable: true }
        scrape_path: { type: string }
        schedule: { type: string }
        enabled: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CategoryCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        scrape_path: { type: string }
        category_id: { type: integer }
        schedule: { type: string }
        enabled: { type: boolean, default: true }
    CategoryUpdate:
      type: object
      properties:
        name: { type: string }
        scrape_path: { type: string }
        category_id: { type: integer }
        schedule: { type: string }
        enabled: { type: boolean }